requirements.txt
playwright==1.48.0
pyyaml==6.0.2
pydantic==2.10.3
discord-webhook==1.3.1
python-dotenv==1.0.1
